% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dolphin}
\alias{dolphin}
\title{Measurements from the vertebral column of a dolphin}
\format{
A matrix with 40 vertebrae and 16 measurements. Column 1, vertebra, is the positional
information.
}
\usage{
dolphin
}
\description{
Linear and angular measurements from Platanista gangetica SMNS 45653
}
\keyword{datasets}
