% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphoRegions-package.R
\docType{package}
\name{MorphoRegions-package}
\alias{MorphoRegions}
\alias{MorphoRegions-package}
\title{MorphoRegions: Analysis of Regionalization Patterns in Serially Homologous Structures}
\description{
Computes the optimal number of regions (or subdivisions) and their position in serial structures without a priori assumptions and to visualize the results. After reducing data dimensionality with the built-in function for data ordination, regions are fitted as segmented linear regressions along the serial structure. Every region boundary position and increasing number of regions are iteratively fitted and the best model (number of regions and boundary positions) is selected with an information criterion. This package expands on the previous 'regions' package (Jones et al. (2018) \doi{10.1126/science.aar3126}) with improved computation and more fitting and plotting options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aagillet.github.io/MorphoRegions/}
  \item \url{https://github.com/aagillet/MorphoRegions}
  \item Report bugs at \url{https://github.com/aagillet/MorphoRegions/issues}
}

}
\author{
\strong{Maintainer}: Amandine Gillet \email{gillet.aman@gmail.com} (\href{https://orcid.org/0000-0001-9763-6061}{ORCID})

Authors:
\itemize{
  \item Amandine Gillet \email{gillet.aman@gmail.com} (\href{https://orcid.org/0000-0001-9763-6061}{ORCID})
  \item Katrina Jones \email{katrina.jones@bristol.ac.uk} (\href{https://orcid.org/0000-0003-1088-0497}{ORCID})
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})
  \item Stephanie Pierce \email{spierce@oeb.harvard.edu} (\href{https://orcid.org/0000-0003-0717-1841}{ORCID})
}

}
\keyword{internal}
