% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions-package.R
\docType{package}
\name{regions-package}
\alias{regions}
\alias{regions-package}
\title{regions: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
