% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions2-package.R
\docType{package}
\name{regions2-package}
\alias{regions2}
\alias{regions2-package}
\title{regions2: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
