% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.regions_pco.R
\name{subset.regions_pco}
\alias{subset.regions_pco}
\title{Subset a \code{regions_pco} object}
\usage{
\method{subset}{regions_pco}(x, subset, drop = TRUE, ...)
}
\arguments{
\item{x}{a \code{regions_pco} object, the output of a call to \code{\link[=svdPCO]{svdPCO()}}.}

\item{subset}{a \code{logical} vector with length equal to the number of rows of the original dataset passed to \code{process_measurements()} before \code{svdPCO()}. The vector is evaluated in the environment of the original dataset.}

\item{drop}{\code{logical}. When \code{TRUE}  (the default), the observations not in the subset are removed from the dataset. When \code{FALSE}, all observations remain, but the indices of the subsetted observations are stored. This only affects how \code{plotvertmap()} displays observations when subsetting changes the range of the vertebra indices. See Examples.}

\item{...}{ignored.}
}
\value{
A \code{regions_pco} object subsetted according to the argument passed to \code{subset}.
}
\description{
\code{subset()} allows one to subset a \code{regions_pco} object, the output of a call to \code{\link[=svdPCO]{svdPCO()}}. This function is useful only to include other observations when computing the PCOs but to focus the analysis on a core set of observations.
}
\examples{
data("alligator")

alligator_data <- process_measurements(alligator,
                                       pos = "Vertebra")

# Compute PCOs
alligator_PCO <- svdPCO(alligator_data)

# Plot vertebrae before subsetting
plotvertmap(alligator_PCO, text = TRUE)

# Subset data after estimating PCOs
alligator_PCO_subset <- subset(alligator_PCO,
                               Vertebra < 20)

# Plot vertebrae after subsetting; vertebrae outside
# subset are dropped
plotvertmap(alligator_PCO_subset, text = TRUE)

# Subset data after estimating PCOs with `drop = FALSE`
alligator_PCO_subset <- subset(alligator_PCO,
                               Vertebra < 20,
                               drop = FALSE)

# Plot vertebrae after subsetting; vertebrae outside
# subset remain and are treated as missing
plotvertmap(alligator_PCO_subset, text = TRUE)
}
\seealso{
\code{\link[=subsample]{subsample()}} for randomly subsetting a \code{regions_pco} object.
}
