% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seal_gmm}
\alias{seal_gmm}
\title{Geometric morphometric data and PC scores of the vertebral column a seal}
\format{
A list containing: a 3D array of Procrustes aligned 3D landmarks (\code{coord_gpa}) of 25 presacral vertebrae with 40 landmarks per vertebra, a matrix of PC scores from a PCA performed on Procrustes aligned landmarks (\code{scores}),
and a numeric vector with the eigenvalues from the PCA (\code{eigenvals}).
}
\usage{
data("seal_gmm")
}
\description{
Procrusted aligned 3d landmarks (\code{coord_gpa}) and PC scores (\code{scores}) of the backbone of Phoca vitulina (ZMUC 160).
}
\references{
Esteban, J. M., Martín-Serra, A., Pérez-Ramos, A., Mulot, B., Jones K. E., & Figueirido, B.  (2023). The impact of the land-to-sea transition on evolutionary integration and modularity of the pinniped backbone. Nature Communications, 6, 1141.
}
\keyword{datasets}
