% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolves_gmm}
\alias{wolves_gmm}
\title{Geometric morphometric data and PC scores of the vertebral column two wolves}
\format{
A list containing: a 3D array of Procrustes aligned 3D landmarks (\code{coord_gpa}) of 10 subsampled vertebrae per specimen, and 30 landmarks per vertebra, a matrix of PC scores from a PCA performed on Procrustes aligned landmarks (\code{scores}),
a numeric vector with the eigenvalues from the PCA (\code{eigenvals}), a numeric vector with the position information of vertebrae (\code{vertebrae}), and a vector with the names of the specimens (\code{specimens}).
}
\usage{
data("wolves_gmm")
}
\description{
Procrustes aligned 3d landmarks (\code{coord_gpa}) and PC scores (\code{scores}) from  10 subsampled vertebrae along the backbone of two specimens of Canis lupus (MU 071 and MU 073).
}
\references{
Schwab, J., Figueirido, B., & Jones, K. E. (2026). Ecological inference from isolated vertebrae: Evaluating functional signal across the carnivoran spine. Journal of Morphology, 287(1), e70109.
}
\keyword{datasets}
